Add EngagementRecipientStep component to streamline recipient selection process
